+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Prism Library Release Notes
 ==========================
 
+###v1.2
+* Added Bootstrap 3 layouts to PrismUI.
+* Added methods getKeys, getParams, setParam to Prism\Database\Table and Prism\Database\TableImmutable.
+
 ###v1.1
 * Removed UI.bootstrap2FileStyle and bootstrap3FileStyle. Replaced with bootstrap3FileInput.
 * Renamed UI.bootstrapFileUploadStyle to UI.bootstrap2FileInput